requirements.txt
pandas
numpy
fastapi
uvicorn
scapy
requests
matplotlib
flask
python-dotenv
openai
